Move paging structures to `Memory.i386.hpp`
