+++ kernel/mem/paging.cpp
@@ -25,7 +25,7 @@ namespace Paging {
 #define PAGE_ENTRIES        1024
 #define ADDRESS_SPACE_SIZE  0x100000000
 #define KADDR_TO_PHYS(addr) ((addr) - KERNEL_BASE)
-#define ADDR(addr)          ((union Address) { .val = (addr) })
+#define ADDR(addr)          ((union Arch::Memory::Address) { .val = (addr) })
 
 static Mutex pagingLock("paging");
 
@@ -36,11 +36,11 @@ static Bitset<size_t, MEM_BITMAP_SIZE> mappedMemory;
 static Bitset<size_t, MEM_BITMAP_SIZE> mappedPages;
 
 static uint32_t pageDirectoryAddress;
-static struct Table* pageDirectoryVirtual[PAGE_ENTRIES];
+static struct Arch::Memory::Table* pageDirectoryVirtual[PAGE_ENTRIES];
 
 // both of these must be page aligned for anything to work right at all
-static struct DirectoryEntry pageDirectoryPhysical[PAGE_ENTRIES] SECTION(".page_tables,\"aw\", @nobits#");
-static struct Table pageTables[PAGE_ENTRIES] SECTION(".page_tables,\"aw\", @nobits#");
+static struct Arch::Memory::DirectoryEntry pageDirectoryPhysical[PAGE_ENTRIES] SECTION(".page_tables,\"aw\", @nobits#");
+static struct Arch::Memory::Table pageTables[PAGE_ENTRIES] SECTION(".page_tables,\"aw\", @nobits#");
 
 // Function prototypes
 static void pageFaultCallback(struct registers* regs);
@@ -49,7 +49,7 @@ static void mapEarlyMem();
 static void mapKernel();
 static uint32_t findNextFreeVirtualAddress(int seq);
 static uint32_t findNextFreePhysicalAddress();
-static inline void mapKernelPageTable(uint32_t idx, struct Table* table);
+static inline void mapKernelPageTable(uint32_t idx, struct Arch::Memory::Table* table);
 static inline void setPageDirectory(uint32_t Directory);
 static void argumentsCallback(const char* arg);
 
@@ -85,7 +85,7 @@ static void pageFaultCallback(struct registers* regs)
     PANIC(regs);
 }
 
-static inline void mapKernelPageTable(uint32_t idx, struct Table* table)
+static inline void mapKernelPageTable(uint32_t idx, struct Arch::Memory::Table* table)
 {
     pageDirectoryVirtual[idx] = table;
     pageDirectoryPhysical[idx] = {
@@ -112,11 +112,11 @@ static void initDirectory()
         mapKernelPageTable(i, &pageTables[i]);
         // clear out the page tables
         for (int j = 0; j < PAGE_ENTRIES; j++) {
-            memset(&pageTables[i].pages[j], 0, sizeof(struct TableEntry));
+            memset(&pageTables[i].pages[j], 0, sizeof(struct Arch::Memory::TableEntry));
         }
     }
     // recursively map the last page table to the page directory
-    mapKernelPageTable(PAGE_ENTRIES - 1, (struct Table*)&pageDirectoryPhysical[0]);
+    mapKernelPageTable(PAGE_ENTRIES - 1, (struct Arch::Memory::Table*)&pageDirectoryPhysical[0]);
     for (uint32_t i = PAGE_ENTRIES * (PAGE_ENTRIES - 1); i < PAGE_ENTRIES * PAGE_ENTRIES; i++) {
         mappedPages.Set(i);
     }
@@ -124,7 +124,7 @@ static void initDirectory()
     pageDirectoryAddress = KADDR_TO_PHYS((uint32_t)&pageDirectoryPhysical[0]);
 }
 
-void mapKernelPage(union Address vaddr, union Address paddr)
+void mapKernelPage(union Arch::Memory::Address vaddr, union Arch::Memory::Address paddr)
 {
     // Set the page directory entry (pde) and page table entry (pte)
     uint32_t pde = vaddr.page.dirIndex;
@@ -133,7 +133,7 @@ void mapKernelPage(union Address vaddr, union Address paddr)
     if (vaddr.page.offset != 0) {
         PANIC("Attempted to map a non-page-aligned virtual address.\n");
     }
-    TableEntry* entry = &(pageTables[pde].pages[pte]);
+    Arch::Memory::TableEntry* entry = &(pageTables[pde].pages[pte]);
     // Print a debug message to serial
     if (is_mapping_output_enabled) {
         debugf("map 0x%08x to 0x%08x, pde = 0x%08x, pte = 0x%08x\n", paddr.val, vaddr.val, pde, pte);
@@ -168,7 +168,7 @@ void mapKernelPage(union Address vaddr, union Address paddr)
 
 void mapKernelRangeVirtual(uintptr_t begin, uintptr_t end)
 {
-    union Address a;
+    union Arch::Memory::Address a;
     for (a = ADDR(begin); a.val < end; a.val += ARCH_PAGE_SIZE) {
         mapKernelPage(a, a);
     }
@@ -176,9 +176,9 @@ void mapKernelRangeVirtual(uintptr_t begin, uintptr_t end)
 
 void mapKernelRangePhysical(uintptr_t begin, uintptr_t end)
 {
-    union Address a;
+    union Arch::Memory::Address a;
     for (a = ADDR(begin); a.val < end; a.val += ARCH_PAGE_SIZE) {
-        union Address phys {
+        union Arch::Memory::Address phys {
             .val = KADDR_TO_PHYS(a.val)
         };
         mapKernelPage(a, phys);
@@ -227,7 +227,7 @@ void* newPage(uint32_t size)
         uint32_t phys_page_idx = findNextFreePhysicalAddress();
         if (phys_page_idx == SIZE_MAX)
             return NULL;
-        union Address phys = {
+        union Arch::Memory::Address phys = {
             .val = phys_page_idx * ARCH_PAGE_SIZE,
         };
         mapKernelPage(ADDR((uint32_t)i * ARCH_PAGE_SIZE), phys);
@@ -244,7 +244,7 @@ void freePage(void* page, uint32_t size)
     for (uint32_t i = page_index; i < page_index + page_count; i++) {
         mappedPages.Reset(i);
         // this is the same as the line above
-        struct TableEntry* pte = &(pageTables[i / PAGE_ENTRIES].pages[i % PAGE_ENTRIES]);
+        struct Arch::Memory::TableEntry* pte = &(pageTables[i / PAGE_ENTRIES].pages[i % PAGE_ENTRIES]);
         // the frame field is actually the page frame's index basically it's frame 0, 1...(2^21-1)
         mappedMemory.Reset(pte->frame);
         // zero it out to unmap it
